pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pharma-csa-framework"
version = "0.1.0"
description = "CSA Validation Automation Framework — Bridging Pharmaceutical Quality & Software Test Automation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Milton Klun"},
]
keywords = ["csa", "fda", "pharma", "validation", "playwright", "testing", "quality"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    # Demo App (System Under Validation)
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Risk Engine & System Inventory
    "pyyaml>=6.0.1",

    # Report Generation
    "jinja2>=3.1.3",

    # Evidence Capture
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    # Test Automation (Scripted)
    "pytest>=8.0.0",
    "pytest-playwright>=0.4.4",
    "pytest-asyncio>=0.23.0",
    "playwright>=1.41.0",

    # Test Reporting
    "allure-pytest>=2.13.2",

    # Code Quality
    "ruff>=0.2.0",
    "mypy>=1.8.0",

    # Coverage
    "pytest-cov>=4.1.0",

    # HTTP testing
    "httpx>=0.26.0",
]

[project.urls]
Homepage = "https://github.com/miltonklun/pharma-csa-framework"
Documentation = "https://github.com/miltonklun/pharma-csa-framework/docs"
Repository = "https://github.com/miltonklun/pharma-csa-framework"
Issues = "https://github.com/miltonklun/pharma-csa-framework/issues"

[tool.pytest.ini_options]
testpaths = ["tests", "test_suites"]
markers = [
    "scripted: Scripted tests for HIGH process risk features (CSA Step 3a)",
    "unscripted: Unscripted/exploratory tests for NOT HIGH process risk features (CSA Step 3b)",
    "audit_trail: Tests validating audit trail compliance (21 CFR Part 11 §11.10(e))",
    "e_signature: Tests validating electronic signature compliance (21 CFR Part 11 §11.50/§11.70)",
    "access_control: Tests validating role-based access control (21 CFR Part 11 §11.10(d))",
    "data_integrity: Tests validating ALCOA+ data integrity (PIC/S PI 041)",
    "high_risk: Tests for HIGH process risk features",
    "not_high_risk: Tests for NOT HIGH process risk features",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "A", "C4", "PT"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.setuptools.packages.find]
include = [
    "demo_app*", 
    "system_inventory*", 
    "risk_engine*", 
    "test_suites*", 
    "evidence_capture*", 
    "report_generator*"
]
